refactor(agent): extract completion card builder
