+++ apps/server/api/src/services/agent-orchestrator/agent-orchestrator.module.ts
@@ -30,6 +30,7 @@ import { AnalyticsModule } from '@api/endpoints/analytics/analytics.module';
 import { MarketplaceIntegrationModule } from '@api/marketplace-integration/marketplace-integration.module';
 import { AgentMessageBusModule } from '@api/services/agent-campaign/agent-message-bus.module';
 import { AgentContextAssemblyModule } from '@api/services/agent-context-assembly/agent-context-assembly.module';
+import { AgentCompletionCardBuilderService } from '@api/services/agent-orchestrator/agent-completion-card-builder.service';
 import { AgentOrchestratorController } from '@api/services/agent-orchestrator/agent-orchestrator.controller';
 import { AgentOrchestratorService } from '@api/services/agent-orchestrator/agent-orchestrator.service';
 import { AgentStreamPublisherModule } from '@api/services/agent-orchestrator/agent-stream-publisher.module';
@@ -94,6 +95,7 @@ import { forwardRef, Module } from '@nestjs/common';
     forwardRef(() => SkillRuntimeModule),
   ],
   providers: [
+    AgentCompletionCardBuilderService,
     AgentOrchestratorService,
     AgentRouteRewriteService,
     AgentToolExecutorService,
